// src/app/manual-case/manual-case.interface.ts
export enum ManualCasePath {
  ManualCase = 'manual-case',
  NewCaseType = 'type',
  EditCaseType = 'edit-type',
  NewCaseDetail = 'case-detail',
  EditCaseDetail = 'edit-case-detail',
  NewProsecutor = 'prosecutor',
  EditProsecutor = 'edit-prosecutor',
  NewHearing = 'hearing',
  EditHearing = 'edit-hearing',
  NewDefendant = 'defendant',
  EditDefendant = 'edit-defendant',
  NewOffence = 'offence',
  EitherWayOffence = 'either-way-offence',
  EditEitherWayOffence = 'edit-either-way-offence',
  EditOffence = 'edit-offence',
  CaseSummary = 'summary',
  CaseCreated = 'case-created/:reference/:caseId',
  DuplicatedProsecutor = 'duplicated-prosecutor'
}